+++ src/components/auth/AuthInterceptor/index.tsx
@@ -1,12 +1,9 @@
 import AuthAxios from '@/apis/@core/authInstance';
-import { refresh } from '@/apis/auth';
-import { AUTH_PATH } from '@/constants/routes';
+import { useRefresh } from '@/hooks/auth/useRefresh';
 import accessTokenAtom from '@/recoil/auth/accessToken';
 import axios from 'axios';
 import { useEffect } from 'react';
-import { useCookies } from 'react-cookie';
-import { useNavigate } from 'react-router';
-import { useRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 
 interface IAuthInterceptorProps {
   children: React.ReactNode;
@@ -21,10 +18,8 @@ interface IAuthInterceptorProps {
  *      2) 401에러가 뜨는 경우 refresh 요청
  */
 const AuthInterceptor = ({ children }: IAuthInterceptorProps) => {
-  const navigate = useNavigate();
-
-  const [accessToken, setAccessToken] = useRecoilState(accessTokenAtom);
-  const [, setCookie] = useCookies(['isLoggedIn']);
+  const accessToken = useRecoilValue(accessTokenAtom);
+  const { mutateAsync } = useRefresh();
 
   const responseInterceptor = AuthAxios.interceptors.request.use((config) => {
     if (accessToken) {
@@ -45,28 +40,11 @@ const AuthInterceptor = ({ children }: IAuthInterceptorProps) => {
       // 권한이 필요한 api 요청에서 access token 에러가 발생하는 경우
       if (status === 401) {
         try {
-          const { accessToken, refreshTokenExpiration } = await refresh();
-          const expires = new Date(refreshTokenExpiration);
-
-          setAccessToken(accessToken);
-          setCookie('isLoggedIn', true, {
-            expires,
-          });
+          const { accessToken } = await mutateAsync();
           config.headers.Authorization = `Bearer ${accessToken}`;
-
+          
           return axios(config);
         } catch (err) {
-          if (axios.isAxiosError(err)) {
-            const { status } = err;
-
-            // 리프레시 요청에서도 에러가 발생하는 경우는 로그인 페이지로 이동
-            if (status === 400 || status === 401) {
-              alert('세션이 만료되었습니다. 다시 로그인해 주세요.');
-              navigate(AUTH_PATH.LOGIN);
-              return;
-            }
-          }
-
           return Promise.reject(err);
         }
       }
